requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
boto3==1.14.57
botocore==1.17.57
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.41.0
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-storages==1.10
django-taggit==1.2.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Markdown==3.2.1
oauthlib==3.1.0
pilkit==2.0
Pillow==7.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
